Change parameter MyStyles to MyTypes by mapping types to their styles
The principle being that this is less work for the client up front
- enumerating the types you want to use with Qplot is less painful than
having to name a set of often arbitrary 'style variants' and then
enumerate these.

This also admits being able to directly map a users types to a list of
styles (a variant), which is cleaner than the indirection behind naming
the variant and then map the users types to these names.

Technically this could lead to more memory for Qplot handling duplicate
variants. One to consider later. See how this goes for now.