Remove `check_tuple` when using WITH_BOOST_VARIANT
This trait makes use of `std::variant_size_v` which is not preset (as is)
on boost. Remains a todo to find an alternative, so just disabling this
check for now.

Note the irony that (at least currently) the code is broken with
boost::variant anyway. Also to do.

Todo todo todo